refactor: added fetch user custom hook
